Couple of related changes:
* Split `CompressedTypedChannel` into `CompressedBincodeChannel` and
  `CompressedTypedChannel` to match the normal reliable channel, allowing for
  protocols that change types.
* Change the integer types of many of the packet and message size limits to be
  `u16`.  There was an implicit assumption in the unreliable channel that the
  max packet size is not larger than a u16, which meant that the maximum message
  length must not be larger than a u16 which is what is used in the
  serialization protocol.  The reliable channels also required maximum message
  lengths to be less than u16 for the same reason.  Rather than have to explain
  this with max value constants, just change the types to reflect this.  In the
  future, maybe a usize maximum and VLQ integer encoding would be better.
* The maximum size of a `CompressedTypeChannel` message was 64k exactly, not
  u16::MAX.  This assumption was strange and mismatched the non-compresed
  reliable channel.  The size sent was the serialized size minus one, which
  can't be violated from a normal implementation of Serialize that writes
  anything of nonzero size, but there's no point in being so magical about it.

+++ src/message_channels.rs
@@ -43,11 +43,11 @@ pub enum MessageChannelMode {
     Unreliable,
     Reliable {
         reliability_settings: reliable_channel::Settings,
-        max_message_len: usize,
+        max_message_len: u16,
     },
     Compressed {
         reliability_settings: reliable_channel::Settings,
-        max_chunk_len: usize,
+        max_chunk_len: u16,
     },
 }
